# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "post": {
        "description": "Crea un registro asociado a la tienda del usuario autenticado. Roles permitidos: OWNER, MANAGER.",
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada (POST devuelve 201).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear producto",
        "tags": [
          "products"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      },
      "get": {
        "description": "Listado paginado por tienda. Busca por título o SKU; permite status y nombre exacto de category. Orden predeterminado createdAt desc. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página solicitada; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Texto de búsqueda, sin distinguir mayúsculas",
            "schema": {
              "example": "cafe",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo del modelo Prisma para ordenar; no usar nombres calculados de la respuesta",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "description": "Dirección del orden",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Estado o rol permitido",
            "schema": {
              "$ref": "#/components/schemas/ProductStatus"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Nombre exacto de categoría, sin distinguir mayúsculas; no es un UUID",
            "schema": {
              "example": "Bebidas",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar productos",
        "tags": [
          "products"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "description": "Obtiene el registro identificado por UUID dentro de la tienda del usuario. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar producto",
        "tags": [
          "products"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      },
      "patch": {
        "description": "Actualización parcial: las propiedades omitidas conservan su valor. Roles permitidos: OWNER, MANAGER.",
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar producto",
        "tags": [
          "products"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      },
      "delete": {
        "description": "Elimina el registro y devuelve sus datos. Las relaciones existentes pueden impedir la eliminación. Roles permitidos: OWNER.",
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado en la tienda del usuario o registro relacionado inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar producto",
        "tags": [
          "products"
        ],
        "x-roles": [
          "OWNER"
        ]
      }
    },
...
  },
  "info": {
    "title": "La Hermandad — API POS",
    "description": "API del punto de venta: autenticación, usuarios, categorías, productos, clientes y pedidos con pagos. Las rutas protegidas obtienen la tienda desde el usuario autenticado. Use login o register y pegue el token en Authorize. Los JWT duran 24 horas. Los cuerpos rechazan propiedades desconocidas. Los importes de entrada son números y los Decimal de salida son cadenas. Todos los POST devuelven 201; PATCH y DELETE devuelven 200 con cuerpo. Los permisos se indican mediante x-roles en cada operación.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Pegue únicamente el token obtenido en auth/login o auth/register."
      }
    },
    "schemas": {
...
      "ProductStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "ACTIVE",
          "ARCHIVED"
        ],
        "description": "Estado o rol permitido"
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Café",
            "minLength": 3
          },
          "description": {
            "type": "string",
            "description": "Descripción",
            "example": "Descripción de ejemplo"
          },
          "sku": {
            "type": "string",
            "description": "Código SKU, único por tienda",
            "example": "CAF-001"
          },
          "price": {
            "type": "number",
            "description": "Precio de venta",
            "example": 25.5,
            "minimum": 0
          },
          "costPrice": {
            "type": "number",
            "description": "Costo del producto",
            "example": 15,
            "minimum": 0
          },
          "compareAtPrice": {
            "type": "number",
            "description": "Precio de comparación",
            "example": 30,
            "minimum": 0
          },
          "trackInventory": {
            "type": "boolean",
            "description": "Controlar existencias",
            "example": true
          },
          "inventoryQuantity": {
            "type": "number",
            "description": "Existencias del producto; enviar un entero",
            "example": 20,
            "minimum": 0
          },
          "status": {
            "description": "Estado o rol permitido",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductStatus"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "Identificador UUID del recurso",
            "example": "123e4567-e89b-42d3-a456-426614174000",
            "format": "uuid"
          }
        },
        "required": [
          "title",
          "price"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Café",
            "minLength": 3
          },
          "description": {
            "type": "string",
            "description": "Descripción",
            "example": "Descripción de ejemplo"
          },
          "sku": {
            "type": "string",
            "description": "Código SKU, único por tienda",
            "example": "CAF-001"
          },
          "price": {
            "type": "number",
            "description": "Precio de venta",
            "example": 25.5,
            "minimum": 0
          },
          "costPrice": {
            "type": "number",
            "description": "Costo del producto",
            "example": 15,
            "minimum": 0
          },
          "compareAtPrice": {
            "type": "number",
            "description": "Precio de comparación",
            "example": 30,
            "minimum": 0
          },
          "trackInventory": {
            "type": "boolean",
            "description": "Controlar existencias",
            "example": true
          },
          "inventoryQuantity": {
            "type": "number",
            "description": "Existencias del producto; enviar un entero",
            "example": 20,
            "minimum": 0
          },
          "status": {
            "description": "Estado o rol permitido",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductStatus"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "Identificador UUID del recurso",
            "example": "123e4567-e89b-42d3-a456-426614174000",
            "format": "uuid"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "description": "Mensaje del servicio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "property unknown should not exist"
            ]
          },
          "error": {
            "type": "string",
            "description": "Nombre HTTP del error, cuando Nest lo incluye",
            "example": "Bad Request"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": [
          "totalItems",
          "currentPage",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ],
        "properties": {
          "totalItems": {
            "type": "integer",
            "example": 1
          },
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "example": 1
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "Product": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "sku",
          "price",
          "costPrice",
          "compareAtPrice",
          "trackInventory",
          "inventoryQuantity",
          "status",
          "category",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Café"
          },
          "description": {
            "type": "string",
            "description": "Descripción",
            "example": "Ejemplo",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "description": "SKU",
            "example": "CAF-001",
            "nullable": true
          },
          "price": {
            "type": "string",
            "description": "Decimal de Prisma serializado como cadena JSON; no garantiza ceros finales.",
            "example": "25.5"
          },
          "costPrice": {
            "type": "string",
            "description": "Decimal de Prisma serializado como cadena JSON; no garantiza ceros finales.",
            "example": "25.5",
            "nullable": true
          },
          "compareAtPrice": {
            "type": "string",
            "description": "Decimal de Prisma serializado como cadena JSON; no garantiza ceros finales.",
            "example": "25.5",
            "nullable": true
          },
          "trackInventory": {
            "type": "boolean",
            "example": true
          },
          "inventoryQuantity": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "ARCHIVED"
            ],
            "example": "DRAFT"
          },
          "category": {
            "type": "object",
            "required": [
              "id",
              "name"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "example": "123e4567-e89b-42d3-a456-426614174000"
              },
              "name": {
                "type": "string",
                "description": "Categoría",
                "example": "Bebidas"
              }
            },
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-27T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-27T12:00:00.000Z"
          }
        }
      },
...
      "ProductPage": {
        "type": "object",
        "required": [
          "meta",
          "results"
        ],
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
...
    }
  }
}